requirements.txt
tensorflow
numpy
flask
pillow
werkzeug
gevent